Use kolla variable to set inspector DHCP config
Fixes an issue with ironic prechecks, which now require
ironic_dnsmasq_dhcp_range to be set when ironic is enabled.

Also uses the new kolla variable ironic_dnsmasq_default_gateway to
configure the gateway for the inspection network.

These changes mean that we can drop our custom override of
ironic-dnsmasq.conf, since it now matches the file in kolla ansible.

This change depends on the Rocky release of Kolla Ansible.

Change-Id: I0b2e094904772fc6e14f1916e20ef207161ec001